requirements.txt
dash
plotly
pandas
numpy
sqlalchemy
psycopg2
python-dotenv
